+++ CHANGELOG.md
@@ -13,6 +13,9 @@
   encoding from locale settings, even if the connection database has a different
   one.
 
+* Fall back to a permissive-but-hopefully-more-robust loader for character type
+  values if client encoding is really `ascii`, when using the psycopg backend.
+
 ## pg\_activity 3.1.1 - 2023-03-06
 
 ### Fixed